checkout old galaxy_net.py
